% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{analise_residuos}
\alias{analise_residuos}
\title{Análise de Resíduos de um Modelo de Regressão}
\usage{
analise_residuos(modelo)
}
\arguments{
\item{modelo}{Uma lista contendo os resultados do ajuste do modelo.
Deve incluir:
\itemize{
  \item \code{coeficientes}: vetor de coeficientes do modelo.
  \item \code{residuos}: vetor de resíduos do modelo.
  \item \code{valores_preditos}: vetor de valores preditos pelo modelo.
}}
}
\value{
Uma lista de gráficos gerados, contendo:
\itemize{
  \item \code{grafico_res_ajustados}: gráfico de resíduos vs. valores ajustados.
  \item \code{grafico_histograma_res}: histograma dos resíduos.
  \item \code{grafico_qq_res}: gráfico Q-Q dos resíduos.
  \item \code{grafico_acf_residuos}: gráfico de autocorrelação dos resíduos.
  \item \code{shapiro_wilks}: resultado do teste de Shapiro-Wilk para testar a normalidade dos resíduos.
}
}
\description{
Esta função gera gráficos para analisar os resíduos de um modelo de regressão linear.
Ela plota o gráfico de resíduos vs. valores ajustados, um histograma dos resíduos,
um gráfico Q-Q para verificar a normalidade dos resíduos, e um gráfico de autocorrelação
dos resíduos.
}
\examples{
# Ajustar o modelo
resultado <- regressao_linear(X, Y)

# Analisar os resíduos
graficos_modelo <- analise_residuos(resultado)

# Exibir gráficos dos resíduos
print(graficos_modelo$grafico_res_ajustados)
print(graficos_modelo$grafico_histograma_res)
print(graficos_modelo$grafico_qq_res)
print(graficos_modelo$grafico_acf_residuos)
}
