% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\docType{data}
\name{human}
\alias{human}
\title{Conjunto de Dados Exemplo}
\format{
Um data frame com 100 observações e 7 variáveis:
\describe{
  \item{idade}{Idade do indivíduo (numérico, entre 18 e 65).}
  \item{renda_anual}{Renda anual do indivíduo (numérico, distribuído conforme uma distribuição gamma).}
  \item{ocupacao}{Tipo de ocupação do indivíduo (categórico: "Empregado", "Desempregado", "Autônomo").}
  \item{sexo}{Sexo do indivíduo (categórico: "Homem", "Mulher").}
  \item{altura}{Altura do indivíduo em centímetros (numérico, distribuído normalmente).}
  \item{anos_exp}{Anos de experiência do indivíduo (numérico, valor entre 0 e 17).}
  \item{satisfacao}{Nível de satisfação do indivíduo (numérico, varia de 0 a 10).}
}
}
\source{
Arquivo gerado aleatoriamente com a função `sample`, `rgamma` e `rnorm`.
}
\usage{
human
}
\description{
Um conjunto de dados simulado contendo informações sobre indivíduos,
incluindo idade, renda anual, ocupação, sexo, altura, anos de experiência
e satisfação. Este conjunto de dados pode ser usado para ajustar modelos
de regressão e realizar análises estatísticas.
}
\keyword{datasets}
