% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{predicao}
\alias{predicao}
\title{Predizer novos valores usando um modelo de regressão ajustado}
\usage{
predicao(modelo, X_novo, intercepto = TRUE)
}
\arguments{
\item{modelo}{Uma lista contendo os resultados do modelo ajustado, como os coeficientes.
Deve ser o resultado de uma chamada à função `regressao_linear`.}

\item{X_novo}{Uma matriz de novas variáveis preditoras para as quais se deseja fazer previsões.
A matriz deve ter o mesmo número de colunas e a mesma ordem que o modelo ajustado.}
}
\value{
Um vetor contendo os valores preditos.
}
\description{
Predizer novos valores usando um modelo de regressão ajustado
}
\examples{

# Ajustar um modelo de regressão linear
modelo_lm <- regressao_linear(X, Y, type = "lm")

# Criar uma nova matriz de preditores para predição
X_novo <- as.matrix(novos_dados[, c("idade", "renda_anual", "altura", "anos_exp")])

# Fazer predições para os novos dados
Y_pred_novo <- predicao(modelo_lm, X_novo)

}
