% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previsao.R
\name{previsao}
\alias{previsao}
\title{Previsão}
\usage{
previsao(coeficientes, new_data)
}
\arguments{
\item{coeficientes}{Um vetor de coeficientes obtidos da função regressao_linear.}

\item{new_data}{Um data frame contendo novos dados para previsão.}
}
\value{
Um vetor de predições.
}
\description{
Função para prever novos valores usando coeficientes de um modelo existente.
}
