% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao.R
\name{regressao}
\alias{regressao}
\title{regressao}
\usage{
regressao(y, X)
}
\arguments{
\item{y}{Vetor de observações da variável resposta.}

\item{X}{Matriz de covariáveis.}
}
\value{
Objeto da classe `regressao`, com estimativas dos betas,
valores preditos, resíduos, e outras informações.
}
\description{
aa
}
\details{
A função `plot` pode ser utilizada para visualizar o gráfico
de valores preditos contra valores observados.

A função `predict` pode ser utilizada para fazer novas predições com base
no modelo ajustado.
}
\examples{
modelo <- regressao(exemploy, exemploX)
modelo
plot(modelo)
novoX <- rbind(c(1, 0.5, 3), c(1, 0.8, 1.5))
predict(modelo, novoX)

}
