% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao_linear.R
\name{regressao_linear}
\alias{regressao_linear}
\title{Regressão Linear}
\usage{
regressao_linear(formula, data)
}
\arguments{
\item{formula}{Uma fórmula representando o modelo de regressão linear.}

\item{data}{Um data frame contendo as variáveis na fórmula.}
}
\value{
Uma lista contendo coeficientes, resíduos, valores preditos e gráficos.
}
\description{
Função para realizar regressão linear manualmente e retornar coeficientes, resíduos e gráficos.
}
\examples{
data(mtcars)
resultado <- regressao_linear_manual(mpg ~ wt + hp, data = mtcars)
plot(resultado$grafico_predito_observado)
}
