% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{regressao_linear}
\alias{regressao_linear}
\title{Ajustar um modelo de regressão linear}
\usage{
regressao_linear(
  X,
  Y,
  type = "lm",
  lambda = 0.5,
  intercepto = TRUE,
  nivel_confianca = 0.95
)
}
\arguments{
\item{X}{Uma matriz de variáveis preditoras (apenas valores numéricos).}

\item{Y}{Um vetor da variável resposta.}

\item{type}{Tipo de modelo a ser ajustado: "lm" para regressão linear simples ou "ridge"
para regressão ridge. O valor padrão é "lm".}

\item{lambda}{Parâmetro de penalização para a regressão ridge. O valor padrão é 0.5.}

\item{intercepto}{Lógica indicando se deve ser adicionado um intercepto à matriz X.
TRUE para incluir o intercepto, FALSE caso contrário (padrão é TRUE).}

\item{nivel_confianca}{Nível de significância a ser utilizado no cálculo do intervalo de
confiança para os coeficientes. O padrão é 0.95.}
}
\value{
Uma lista contendo:
\item{coeficientes}{Coeficientes estimados do modelo.}
\item{residuos}{Resíduos do ajuste.}
\item{valores_preditos}{Valores preditos pelo modelo.}
\item{obs_vs_pred}{Gráfico de valores preditos vs. valores observados.}
\item{IC}{Intervalo de confiança dos coeficientes calculados.}
\item{R2}{Coeficiente de determinação do modelo.}
\item{teste_F}{Resultados do teste F, incluindo estatística F e valor p.}
}
\description{
Esta função ajusta um modelo de regressão linear ou ridge a um conjunto de dados, permitindo a
análise e previsão de uma variável resposta com base em variáveis preditoras.
}
\examples{
# Criar a matriz de preditores e vetor de respostas
X <- as.matrix(human[, c("idade", "renda_anual", "altura", "anos_exp")])
Y <- dados$satisfacao

# Ajustar um modelo de regressão linear
modelo_lm <- regressao_linear(X, Y, type = "lm", intercepto = TRUE)

# Ajustar um modelo de regressão ridge
modelo_ridge <- regressao_linear(X, Y, type = "ridge", lambda = 0.5)

# Exibir os coeficientes
print("Coeficientes do modelo linear:")
print(modelo_lm$coeficientes)
print("Coeficientes do modelo ridge:")
print(modelo_ridge$coeficientes)
}
