% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{regressao_linear}
\alias{regressao_linear}
\title{Ajustar um modelo de regressão linear}
\usage{
regressao_linear(X, Y, type = "lm", lambda = 0.5)
}
\arguments{
\item{X}{Uma matriz de variáveis preditoras (incluindo intercepto)}

\item{Y}{Um vetor da variável resposta}

\item{type}{Tipo de modelo a ser ajustado: "lm" para regressão linear simples ou "ridge" para regressão ridge. (default é "lm")
caso seja ridge, há uma normalização nos dados, ou seja, subtrai o desvio padrão e divide pela média.}

\item{lambda}{Parâmetro de penalização para regressão ridge (default é 0.5).}
}
\value{
Uma lista contendo os coeficientes, resíduos, valores preditos e um gráfico.
}
\description{
Ajustar um modelo de regressão linear
}
\examples{
# Criar a matriz de preditores e vetor de respostas
X <- as.matrix(dados[, c("idade", "renda_anual", "altura", "anos_exp")])
X <- cbind(1, X)  # Adiciona uma coluna de 1 para o intercepto
Y <- dados$satisfacao

# Ajustar um modelo de regressão linear
modelo_lm <- regressao_linear(X, Y, type = "lm")

# Ajustar um modelo de regressão ridge
modelo_ridge <- regressao_linear(X, Y, type = "ridge", lambda = 0.5)

# Exibir os coeficientes
print("Coeficientes do modelo linear:")
print(modelo_lm$coeficientes)
print("Coeficientes do modelo ridge:")
print(modelo_ridge$coeficientes)
}
