% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{regressao_linear}
\alias{regressao_linear}
\title{Ajustar um modelo de regressão linear}
\usage{
regressao_linear(
  X,
  Y,
  type = "lm",
  lambda = 0.5,
  intercepto = TRUE,
  nivel_confianca = 0.95
)
}
\arguments{
\item{X}{Uma matriz de variáveis preditoras (apenas valores numéricos).}

\item{Y}{Um vetor da variável resposta.}

\item{type}{Tipo de modelo a ser ajustado: "lm" para regressão linear simples ou "ridge"
para regressão ridge. O valor padrão é "lm".}

\item{lambda}{Parâmetro de penalização para a regressão Ridge. O valor padrão é 0.5, mas
este parâmetro só é utilizado se `type = "ridge"`.}

\item{intercepto}{Lógico, indicando se deve ser adicionado um intercepto à matriz X.
O padrão é TRUE, ou seja, o modelo ajustará um termo constante (intercepto).}

\item{nivel_confianca}{Nível de significância a ser utilizado no cálculo do intervalo de
confiança para os coeficientes. O padrão é 0.95.}
}
\value{
Uma lista contendo:
\item{coeficientes}{Os coeficientes estimados pelo modelo ajustado.}
\item{residuos}{O vetor de resíduos do modelo.}
\item{valores_preditos}{Valores preditos pelo modelo com base em `X`.}
\item{obs_vs_pred}{Gráfico de valores preditos vs. valores observados.}
\item{IC}{Intervalos de confiança dos coeficientes, com base no nível de confiança especificado.}
\item{R2}{Coeficiente de determinação (R²) para o modelo ajustado.}
\item{teste_F}{Resultados do teste F para o modelo, incluindo estatística F e valor-p.}
\item{intercepto}{Indica se o intercepto foi incluído no modelo.}
}
\description{
Esta função ajusta um modelo de regressão linear ou ridge a um conjunto de dados, permitindo a
análise e previsão de uma variável resposta com base em variáveis preditoras.
}
\examples{
# Criar a matriz de preditores e vetor de respostas
X <- as.matrix(human[, c("idade", "renda_anual", "altura", "anos_exp")])
Y <- human$satisfacao

# Ajustar um modelo de regressão linear
modelo_lm <- regressao_linear(X, Y, type = "lm", intercepto = TRUE)

# Ajustar um modelo de regressão ridge com lambda = 0.5
modelo_ridge <- regressao_linear(X, Y, type = "ridge", lambda = 0.5)

# Exibir os coeficientes dos modelos ajustados
print(modelo_lm$coeficientes)
print(modelo_ridge$coeficientes)
}
